Match-id-801bbdf9def5b3d649e5e049d1253021bafa32e1

+++ src/pub_inc/externalinput_pub.h
@@ -96,7 +96,7 @@ const std::string& GetExternalInputProfilingOption();
 
 const std::string& GetExternalInputCannVersion();
 
-const s32& GetExternalInputHcclExecTimeOut();
+const double& GetExternalInputHcclExecTimeOut();
 
 const s32& GetExternalInputHcclLinkTimeOut();
